src/requirements.txt
black~=22.0
flake8>=3.7.9, <5.0
ipython>=7.31.1, <8.0; python_version < '3.8'
ipython~=8.10; python_version >= '3.8'
isort~=5.0
jupyter~=1.0
jupyterlab_server>=2.11.1, <2.16.0
jupyterlab~=3.0
kedro~=0.18.11
kedro-datasets[pandas.CSVDataSet, pandas.ExcelDataSet, pandas.ParquetDataSet]~=1.0
kedro-telemetry~=0.2.0
kedro-viz~=6.0
nbstripout~=0.4
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
pytest~=7.2
scikit-learn~=1.0
numpy~=1.24
pandas~=1.5
pydicom~=2.3
pillow~=9.2
scikit-image~=0.21
scipy~=1.10
torch~=2.0
torchvision~=0.14
#timm~=0.9.2
timm~=0.5.4
tensorboard~=2.13.0
h5py~=3.8
